Make picking tests support new controller API
Controller reform changes the picker to call onPick and onNoPick every frame instead of only when the pick list changes. Modified picking tests to account for this and to render the scene once first before counting frames. Also added new picking test for GVRBoundsPicker.
